+++ packages/form-table/src/public-types.ts
@@ -1,7 +1,10 @@
 import type { Component, PluginObject } from 'vue'
+import type { FormTableComponent, TableRow } from './types.public'
 
 export * from './types.public'
 export { defineFormTableColumns } from './defineFormTableColumns'
+
+export declare function createFormTable<TRow extends TableRow = TableRow>(): FormTableComponent<TRow>
 
 export declare const FormTable: Component
 declare const _default: Component